- Unicode support finished

+++ src/common/debug/PerformanceTest.h
@@ -11,7 +11,7 @@
 #include "Timer.h"
 
 #define TIMER_START(name) { common::Timer::newTimer(#name);
-#define TIMER_N_START(name, nTimes) common::Timer::newTimer(#name); for(int x_##name = 0 ; x_##name < (nTimes) ; x_##name ++ ) {
+#define TIMER_N_START(name, nTimes) common::Timer::newTimer(#name); for(uint64_t x_##name = 0 ; x_##name < (nTimes) ; x_##name ++ ) {
 #define TIMER_END(name) } common::Timer::printResult(#name);
 
 #endif	/* PERFORMANCETEST_H */
